chore: installs redux dev tools and restructures project for redux enhancements
