+++ src/index.js
@@ -6,10 +6,13 @@ import { createStore } from "redux";
 
 import "./index.css";
 import App from "./App";
-import reducer from "./store/reducer";
+import reducer from "./store/reducers/burgerBuilder";
 import registerServiceWorker from "./registerServiceWorker";
 
-const store = createStore(reducer);
+const store = createStore(
+  reducer,
+  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+);
 
 const app = (
   <Provider store={store}>
